feat: Use module.hot instead of react-hot-loader
